refactors test_core_rabbit
